requirements.txt
flask&gt;=2.0.0 

Werkzeug&gt;=2.0.0 

torch&gt;=1.9.0 

torchvision&gt;=0.10.0 

pillow&gt;=8.0.0 

tqdm&gt;=4.0.0 

numpy&gt;=1.19.0